requirements.txt
asgiref==3.4.1
attrs==21.4.0
Automat==20.2.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
constantly==15.1.0
coverage==3.6
cryptography==37.0.4
cssselect==1.1.0
Django==3.2.7
django-cors-headers==3.12.0
django-minio-backend==3.3.2
django-mysql==4.7.0
django-redis==5.1.0
django-restframework==0.0.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.7.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.6.0
itemloaders==1.0.4
jmespath==1.0.1
lxml==4.9.1
minio==7.1.13
mysqlclient==2.1.1
parsel==1.6.0
pika==1.3.0
Protego==0.2.1
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
PyJWT==2.1.0
pyOpenSSL==22.0.0
pytz==2021.1
queuelib==1.6.2
redis==3.5.3
requests==2.28.1
requests-file==1.5.1
Scrapy==2.6.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tldextract==3.3.1
Twisted==22.4.0
twisted-iocpsupport==1.0.2
typing_extensions==4.3.0
urllib3==1.26.10
w3lib==1.22.0
zope.interface==5.4.0